requirements.txt
curl_cffi>=0.6.2
python-dotenv>=1.0.0
typing-extensions>=4.0.0
